Allow users to specify preferred listen address
When listening for IPFS swarm, or peerings, we listen on 0.0.0.0, which
will leave us trying to listen on any/all interfaces available. In
various places though we try and choose an interface to use but it isn't
always the one we want.

By specifying BACALHAU_PREFERRED_ADDRESS it will pick that address over
any others for the IPFS swarm and the peering information.