Add a node selection algorithm to target all nodes
Currently our algorithm for selecting and retrying nodes will select
a number of nodes from an appropriate set until a user-defined number
of successful executions is reached. This is useful for targeting data
processing jobs against a wide and disparate network of different
nodes, as is the common use case for the public Bacalhau network.

We want to meet another use case, which is for a job to be able to
target all nodes that match a given criteria, regardless of how many
nodes that is. This is useful in private networks for running jobs
against a homogeneous set of nodes (i.e. all nodes with some attached
sensor or that are storing the same types of logs).

To achieve this, we split out all of the node targeting logic from the
scheduler and create two new modules – one for the "target any nodes"
case and one for the "target all nodes" case. The scheduler is then
abstracted from the exact logic by which nodes are picked, and just
manages the executions and job state accordingly.

This introduces a new field into the job spec called "TargetAll": a
boolean that switches between the node selection modes (something more
complex may be appropriate in the future that also includes the fields
currently in the Deal struct which are only appropriate for the "any"
case). The default value in old job specs is "any", matching the old
default behaviour. There is currently no way to trigger "all" jobs
from the CLI and this will be added in a later commit.