Exploring a KV based objectstore
This commit is an exploratory one investigating how we might be able to
implement a local objectstore with a KV store before implementing a
distributed objectstore on top of the same KV interface.  So far it
appears to work reasonably well, even allowing us to register callbacks
during a save so that we can extract fields to store in other
'index-like' buckets.

Treating the data provided as opaque bytes is useful, but may introduce
problems with our 'indexes' when we move to a distributed store (such as
etcd).  The issue arises because it's possible our distributed store
will replicate state at the key/value level and won't recognise that our
index buckets (intended to be a lookup for specific object fields)
contain a list. This might push us into a LWW (Last Write Wins) situation
with the indexes, rendering them useless, where what we probably want is
some sort of CRDT convergence on an eventual state.